feat: add a swapAxes option and swap the default axes (#6)
BREAKING CHANGE:

In previous versions, the X axis was along the rows of the data matrix
and the Y axis along the columns. This was not intuitive regarding how
the contours are then drawn. The default was changed to now consider X
along the columns and Y along the rows respectively.
Pass `swapAxes: true` to the Conrec constructor to get the previous
behavior.